requirements.txt
bs4
lxml
matplotlib
pandas
pytest
pytest-cov
pyyaml
